+++ src/main/java/com/callog/callog_diet/domain/entity/Meal.java
@@ -29,17 +29,17 @@ public class Meal {
     private String foodName;
 
     @Column(nullable = false)
-    private Long amount;
+    private Double amount;
 
     @Column
-    private double carbohydrate;
+    private Double carbohydrate;
 
     @Column
-    private double protein;
+    private Double protein;
 
     @Column
-    private double fat;
+    private Double fat;
 
     @Column
-    private double kcal;
+    private Double kcal;
 }
